Do our own IDN logic on the JVM
This adds a dependency on the ASCII-data IDN mapping table originally
written for Kotlin/JS and multiplatform.
