+++ okhttp/src/jvmTest/java/okhttp3/URLConnectionTest.kt
@@ -3039,7 +3039,7 @@ class URLConnectionTest {
     try {
       getResponse(Request("http://-/foo.html".toHttpUrl()))
       fail<Any>()
-    } catch (expected: UnknownHostException) {
+    } catch (expected: IOException) {
     }
   }
