+++ okhttp/src/nonJvmMain/kotlin/okhttp3/internal/-NormalizeNonJvm.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2012 The Android Open Source Project
+ * Copyright (C) 2022 Square, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,13 +15,7 @@
  */
 package okhttp3.internal
 
-import java.net.IDN
-import java.util.Locale
+import com.squareup.okhttpicu.SYSTEM_NORMALIZER
 
-internal actual fun idnToAscii(host: String): String? {
-  return try {
-    IDN.toASCII(host).lowercase(Locale.US)
-  } catch (_: IllegalArgumentException) {
-    null
-  }
-}
+internal actual fun normalizeNfc(string: String): String =
+  SYSTEM_NORMALIZER.normalizeNfc(string)
